changes for manager zone

+++ app/seatbookingui/webapp/ext/controllers/managerZone.js
@@ -43,6 +43,11 @@ sap.ui.define(
                 return Currentdata.value;
         }
 
+        function _refreshExistingModel(reference) {
+            var currrentData = reference._controller.getView().getModel("modelData").getData();
+            reference._controller.getView().getModel("ExistingData").setData(JSON.parse(JSON.stringify(currrentData)));
+        }
+
         function _getEmployeeName() {
             var Currentdata = [];
             $.get({
@@ -72,6 +77,36 @@ sap.ui.define(
             if (Currentdata.value[0])
                 return Currentdata.value[0];
         }
+
+        function fetchSeatData(oContext, reference)
+        {
+             var Currentdata;
+          var oModelref = reference._controller.getView().getModel("modelData");
+          var oModelrefExistingData = reference._controller.getView().getModel("ExistingData");
+          var modelData = oModelref.getData();
+
+          if (modelData.Seats.length === 0){
+
+           var sFilterQuery = `teamID eq '${modelData.Property.teamID}'`;
+            $.get({
+                url: "/seat-booking/TeamSeatMapping",
+                data: {
+                    $filter: sFilterQuery
+                },
+                success: function (data) {
+                    // csrfToken = xhr.getResponseHeader("x-csrf-token");
+                    Currentdata = data;
+                }, async: false
+            });
+            if (Currentdata.value){
+              //  return Currentdata.value[0];
+            oModelref.setProperty("/Seats", Currentdata.value);
+            oModelrefExistingData.setProperty("/Seats", JSON.parse(JSON.stringify(Currentdata.value)));
+            }
+        }
+    }
+
+
 
         return {
             ManagerBtnHandling: function (oContext) {
@@ -108,22 +143,11 @@ sap.ui.define(
                                 "AddEmpBtn": true,
                                 "DeleteEmpBtn": false,
                                 "DeleteEmpBtnTT": "Remove Employee",
-                                "Sync": true
+                                "Sync": true,
+                                "teamID": userData.teamID,
                             },
                             "empList": [],
-                            "Seats": [{
-                                "SeatNo": "BLR04-2F-001",
-                                "Monitors": "2",
-                                "others": "Extension"
-                            }, {
-                                "SeatNo": "BLR04-2F-002",
-                                "Monitors": "1",
-                                "others": "Extension"
-                            }, {
-                                "SeatNo": "BLR04-2F-003",
-                                "Monitors": "2",
-                                "others": "Extension"
-                            }]
+                            "Seats": []
                         };
 
                         var Currentdata = [];
@@ -169,6 +193,9 @@ sap.ui.define(
                         var oEmpModel = new sap.ui.model.json.JSONModel(Newdata);
                         oView.setModel(oEmpModel, "modelData");
 
+                        var oExistingModel = new sap.ui.model.json.JSONModel(JSON.parse(JSON.stringify(Newdata)));
+                        oView.setModel(oExistingModel, "ExistingData");
+
                         //	sap.m.MessageToast.show("Popup to add/update team member, Update equipment on seats, assign different roles to user");
 
                         if (!this._oDialogonManagerZone) {
@@ -209,7 +236,7 @@ sap.ui.define(
             onEmpListSelectionChange: function (oContext) {
                 var oModelref = oContext.getSource().getModel("modelData");
                 var selectedData = oModelref.getProperty(oContext.getSource().getSelectedContextPaths()[0]);
-                if (selectedData.role_roleCode === '01') {
+                if (selectedData.role_roleCode === '01' && selectedData.empIDEdit === false) {
                     oModelref.setProperty("/Property/DeleteEmpBtn", false);
                     oModelref.setProperty("/Property/DeleteEmpBtnTT", 'Manager cannot be removed from this screen');
                 }
@@ -224,38 +251,46 @@ sap.ui.define(
             },
 
             DeleteEmp: function (oContext) {
+                var that = this;
                 var succssFlag;
                 if (oContext.getSource().getParent().getParent().getSelectedContextPaths()) {
                     var paths = oContext.getSource().getParent().getParent().getSelectedContextPaths();
                     if (paths[0]) {
                         var oModelref = oContext.getSource().getModel("modelData");
                         var DeleteIndex = paths[0].split("/").pop();
-
                         var deleteData = oModelref.getData().empList[DeleteIndex];
 
-                        $.ajax({
-                            url: "/seat-booking/TeamEmployeeMaster/" + deleteData.employeeID_ID,
-                            type: 'DELETE',
-                            headers: {
-                                //   "x-csrf-token": csrfToken,
-                                "Content-Type": "application/json"
-                            },
-                            //  data: JSON.stringify(JSON.parse(JSON.stringify(empData))),
-                            success: function (data) {
-                                succssFlag = true;
-                            }, async: false
+                        var existingEmpData = this._controller.getView().getModel('ExistingData').getData().empList;
+
+                        var EmpExist = existingEmpData.find(function (emp) {
+                            return (emp.employeeID_ID === deleteData.employeeID_ID);
                         });
-                     //   if (succssFlag) {
-                            oModelref.getData().empList.splice(DeleteIndex, 1);
-                            oContext.getSource().getParent().getParent().removeSelections();
-                            oModelref.setProperty("/Property/DeleteEmpBtn", false);
-                            this._controller.getView().getModel("modelData").refresh();
-                            sap.m.MessageToast.show("Employee removed!");
-                   //     }
-                   //     else {
-                        //    this._controller.getView().getModel("modelData").refresh();
-                        //    sap.m.MessageToast.show("Please try again");
-                     //   }
+
+                        if (EmpExist) {
+
+                            $.ajax({
+                                url: "/seat-booking/TeamEmployeeMaster/" + deleteData.employeeID_ID,
+                                type: 'DELETE',
+                                headers: {
+                                    //   "x-csrf-token": csrfToken,
+                                    "Content-Type": "application/json"
+                                },
+                                //  data: JSON.stringify(JSON.parse(JSON.stringify(empData))),
+                                success: function (data) {
+                                    succssFlag = true;
+                                    _refreshExistingModel(that);
+                                },
+                                error: function (data) {
+                                    succssFlag = false;
+                                }, async: false
+                            });
+                        }
+
+                        oModelref.getData().empList.splice(DeleteIndex, 1);
+                        oContext.getSource().getParent().getParent().removeSelections();
+                        oModelref.setProperty("/Property/DeleteEmpBtn", false);
+                        this._controller.getView().getModel("modelData").refresh();
+                        sap.m.MessageToast.show("Employee removed!");
 
                     }
                 }
@@ -274,7 +309,7 @@ sap.ui.define(
 
             SaveEmployee: function () {
                 var empData = {}, teamID, empList = [];
-
+                 var successFlag, EmpExist, that = this;
 
                 var sFilterQuery = `employeeID_ID eq '${sap.ushell.Container.getService("UserInfo").getId()}'`;
                 var csrfToken, Currentdata = [];
@@ -298,6 +333,9 @@ sap.ui.define(
                     }
                 }
                 var modelRef = this._controller.getView().getModel('modelData');
+                var existingModelRef = this._controller.getView().getModel('ExistingData');
+                var existingEmpData = existingModelRef.getData().empList;
+               
 
                 modelRef.getData().empList.forEach(element => {
                     empData.employeeID_ID = element.employeeID_ID.toUpperCase();
@@ -306,6 +344,48 @@ sap.ui.define(
 
                     // empList.push(JSON.parse(JSON.stringify(empData)));
 
+                    var EmpExist = existingEmpData.find(function (emp) {
+                        return (emp.employeeID_ID === empData.employeeID_ID);
+                    });
+
+                    if (EmpExist) {
+                        if (EmpExist.role_roleCode === empData.role_roleCode) {
+                            EmpExist = undefined;
+                            return;
+                        }
+                    }
+                    else {
+
+                        if (_getEmployeeNameByID(empData.employeeID_ID)) {
+
+                            sFilterQuery = `employeeID_ID eq '${empData.employeeID_ID}'`;
+                            $.get({
+                                url: "/seat-booking/TeamEmployeeMaster",
+
+                                data: {
+                                    $filter: sFilterQuery
+                                },
+                                success: function (data, status, xhr) {
+                                    Currentdata = data;
+                                }, async: false
+                            });
+                            if (Currentdata.value) {
+                                if (Currentdata.value[0]) {
+                                    if (Currentdata.value[0].teamID !== empData.teamID) {
+                                        successFlag = "Fail";
+                                        sap.m.MessageToast.show(`Employee '${empData.employeeID_ID}' already added for team '${Currentdata.value[0].teamID}'`);
+                                        return;
+                                       
+                                    }
+                                }
+                            }
+                        }
+                        else {
+                             sap.m.MessageToast.show(` '${empData.employeeID_ID}'ID is not correct & will not be saved`);
+                          successFlag = "Fail";
+                             return;
+                        }
+                    }
 
                     $.ajax({
                         url: "/seat-booking/TeamEmployeeMaster/" + empData.employeeID_ID,
@@ -316,17 +396,76 @@ sap.ui.define(
                         },
                         data: JSON.stringify(JSON.parse(JSON.stringify(empData))),
                         success: function (data) {
+                            _refreshExistingModel(that);
+                        }, error: function (data, status) {
+                            const newLocal = 'fail';
+                            successFlag = newLocal;
                         }, async: false
                     });
                 });
-
-
+                if ( successFlag === "Fail") {
+                //    sap.m.MessageToast.show("One of more data is not saved successfully");
+                }
+                else {
+                    sap.m.MessageToast.show("Data saved successfully");
+                }
 
             },
 
             CancelManagerBooking: function () {
                 this._oDialogonManagerZone.close();
             },
+
+            SaveSeat : function (oContext) {
+              var  successFlag, that = this;
+              var modelRef = this._controller.getView().getModel('modelData');
+              var existingModelRef = this._controller.getView().getModel('ExistingData');
+              var existingSeatData = existingModelRef.getData().Seats;       
+
+
+                modelRef.getData().Seats.forEach(element => {
+
+                    var seatExist = existingSeatData.find(function (seat) {
+                        return (seat.seatID === element.seatID
+                                && seat.monitorCount === element.monitorCount
+                               && seat.facility1 === element.facility1
+                               && seat.facility2 === element.facility2 );
+                    });
+
+                    if (seatExist) 
+                        return;
+                    else {
+                        element.facility1 = parseInt(element.facility1);
+                        element.facility2 = parseInt(element.facility2);
+                        element.monitorCount = parseInt(element.monitorCount);
+                           $.ajax({
+                        url: "/seat-booking/TeamSeatMapping/" + element.seatID,
+                        type: 'PUT',
+                        headers: {
+                         //   "x-csrf-token": csrfToken,
+                            "Content-Type": "application/json"
+                        },
+                        data: JSON.stringify(JSON.parse(JSON.stringify(element))),
+                        success: function (data) {
+                            _refreshExistingModel(that);
+                            successFlag = 'success'
+                        }, error: function (data, status) {
+                            const newLocal = 'fail';
+                            successFlag = newLocal;
+                        }, async: false
+                    });
+                    }
+                })
+                if ( successFlag === "fail") {
+                //    sap.m.MessageToast.show("One of more data is not saved successfully");
+                }
+                else {
+                    sap.m.MessageToast.show("Data saved successfully");
+                }
+
+
+            },
+
             ManagerTypeSelection: function (oContext) {
 
                 var modelRef = this._controller.getView().getModel('modelData');
@@ -341,6 +480,7 @@ sap.ui.define(
                     modelRef.setProperty("/Property/EmployeeList", false);
                     modelRef.setProperty("/Property/SeatDetails", true);
                     modelRef.setProperty("/Property/usageAnalytics", false);
+                    fetchSeatData(oContext, this);
                 } else if (index === 2) {
                     modelRef.setProperty("/Property/EmployeeList", false);
                     modelRef.setProperty("/Property/SeatDetails", false);
